update color background
The background color in the class Label was hardcoded. Use the value that is actually set on the attribute.
Adjust the position of the background box so that the text is centered in the box.
Provide a grey background box for fastem ROAs, so that the ROA size is readable always against gray-scale images.
Text color stays white.